Fix featurize_dataframe for features that are matrices
Notes:
1. Now _requires_ output of `featurize` to be a list. Previous
versions of this class would allow classes that produce only
one feature to return a float
2. @computron: `assign` is likely faster than `join` because assign
requires that the new columns be in the same order as the dataframe
and have the same length. Join, on the other hand, is more lenient.
While this isn't proof that `assign` is faster, the fact that it is
less flexible (and we do not need the flexibility) suggests that it
is the right choice. If you'd lke, I can test this.