Fix: Make sure DataFrame indices are copied over
Fixes a problem with `featurize_dataframe`

The current implementation leads to features not being matched
to the proper entry because the indicies of the dataframe
are not passed to the new series containing each feature.
It only works when the first entry is index 0, the second is
index 1, and so on.

I added a test case that fails without the fix, and added
the indices of the dataframe to the new Series being generated