Added a class for multiple featurizers as the same class
The main use case for this class is when featurizers are passed
to an application. This way, that application need not write
functions that loop over all featurizers in a list or set or have
logic that is different for when a user passes a single featurizer
or a list of featurizers.